moved installing alpine req code to vscode package
